+++ CHANGELOG.md
@@ -10,3 +10,7 @@ All notable changes to this project will be documented in this file. See [commit
 * add version configuration file with commit and compare URL formats ([d1e84c4](https://github.com/bocan/codex/commit/d1e84c4fdbd111afb43a2db5aa34afe7aeb96e70))
 
 ## [1.0.2](https://github.com/bocan/ultraformat/compare/v1.0.1...v1.0.2) (2026-03-09)
+
+### Features
+
+* add version tracking and release scripts